+++ src/components/Chat.js
@@ -31,6 +31,11 @@ export const Chat = themed(
       /** Theme object */
       style: PropTypes.object,
       offlineSync: PropTypes.bool,
+      logger: PropTypes.func,
+    };
+
+    static defaultProps = {
+      logger: () => {},
     };
 
     constructor(props) {
@@ -71,7 +76,29 @@ export const Chat = themed(
       this._unmounted = false;
     }
 
+    componentDidMount() {
+      this.props.logger('Chat component', 'componentDidMount', {
+        tags: ['lifecycle', 'chat'],
+        props: this.props,
+        state: this.state,
+      });
+    }
+
+    componentDidUpdate() {
+      this.props.logger('Chat component', 'componentDidUpdate', {
+        tags: ['lifecycle', 'chat'],
+        props: this.props,
+        state: this.state,
+      });
+    }
+
     componentWillUnmount() {
+      this.props.logger('Chat component', 'componentWillUnmount', {
+        tags: ['lifecycle', 'chat'],
+        props: this.props,
+        state: this.state,
+      });
+
       this._unmounted = true;
       this.props.client.off('connection.recovered');
       this.props.client.off('connection.changed');
@@ -113,10 +140,13 @@ export const Chat = themed(
       );
     };
 
-    setActiveChannel = (channel, e) => {
-      if (e !== undefined && e.preventDefault) {
-        e.preventDefault();
-      }
+    setActiveChannel = (channel) => {
+      this.props.logger('Chat component', 'setActiveChannel', {
+        tags: ['chat'],
+        props: this.props,
+        state: this.state,
+      });
+
       if (this._unmounted) return;
       this.setState(() => ({
         channel,
@@ -131,9 +161,15 @@ export const Chat = themed(
       connectionRecovering: this.state.connectionRecovering,
       storage: this.props.storage,
       offlineSync: this.props.offlineSync,
+      logger: this.props.logger,
     });
 
     render() {
+      this.props.logger('Chat component', 'Rerendering', {
+        props: this.props,
+        state: this.state,
+      });
+
       return (
         <ChatContext.Provider value={this.getContext()}>
           {this.props.children}