+++ src/components/index.js
@@ -13,3 +13,4 @@ export { IconBadge } from './IconBadge';
 export { ReactionPicker } from './ReactionPicker';
 export { ReactionPickerWrapper } from './ReactionPickerWrapper';
 export { SendButton } from './SendButton';
+export { FileIcon } from './FileIcon';