+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.3.8] 2019-11-06
+
+- Adding logs to Chat, ChannelList and Channel component
+- Optimizing MessageList and ChannelList component - https://github.com/GetStream/stream-chat-react-native/pull/84
+
+## [0.3.7] 2019-11-04
+
+- Adding support for `AttachmentFileIcon` prop.
+
 ## [0.3.6] 2019-11-04
 
 - Adding support for `actionSheetStyles` prop, so as to add more customizability for styles of actionsheet.